fix: vulnerability of missing fields
As the code was, an adversary could manualy craft an item with missing
field and submit it to the list. The handler would detect the missing field
and no-op. This would cause the item to never show up in UIs that use
the subgraph.

We work around this issue by setting the value manually if it is missing.

Note that an empty string is considered null and caused a crash of the handler
when attempting to save the entity, so we use a hifen instead.